Positionnement complet des portes au commencement de la partie et quelques destructeurs #8

+++ sources/Classes projet/Info.h
@@ -4,7 +4,7 @@ struct InfoObjet {
 	unsigned int ID;
 	Vecteur3d position;
 	Vecteur3d direction;
-	int rotation;
+	double rotation;
 	char* cheminModele;
 	char* cheminTexture;
